I figured out why the code was crashing. The old simulation had damage disabled for body 1. This was the body that was being accelerated into oblivion by the hourglassing force.
What happened is the following: The body got spread out over an enormous range.
Some particles were thousands of mm away from the others. However, since damage
was disabled, these particles were still able to contact stuff. When we called
the contact function, we got an enormous number of buckets. Like on the order
of a billion buckets. For reference, a billion buckets would require about
16 gb just to declare. Thus, the problem was stemming from the contact function.
Enabling damage fixed this issue.

However, going through this process revealed a flaw in the contact algorithm: It
didn't skip damaged particles, even though these shouldn't be able to contact
anything. I added a few lines throughout the contact algorithm which basically
ignore damaged particles (which is what should happen). This change improves
runtime AND CORRECTNESS for simulations which result in damaged particles.

I also bumped the bucket size up to 4 times the contact distance (the 4 is
arbitrary)... I need to play around with this parameter to get optimal performance.

I also reformatted a few things in update_P, and fixed a bunch of typos in Damage.cc