I changed the way that dot, cross, and dyadic products work to reduce the number of conditional branches + function calls.
I ran cache grind on the code. This revealed that the dyadic product
function used a lotttttt of conditional branches. While very few of
these resulted in misses, I decided to take a closer look at
Dyadic_Product. It was using an unrolled loop which and did not directly
access the tensor/vector arrays (it used the [] operators instead).
I fixed all of this. To do so, however, I had to make Dyadic_Product a
friend of the Vector and Tensor classes. I noticed that Dot_Product and
Cross_Product (in the tensor class) suffered from the same issues. I
fixed these as well.

These changes reduced the number of conditional branches (for a
"bouncing simulation") by about 30%. Though this didn't really change
the number of branch misses, it did cut down on calls to the [] operator,
and eliminated a needless loop. I consider that a win.

I haven't measured runtime differences, though I should do that.